# openapi/openapi.json
{
  "components": {
    "parameters": {
...
      "ApplicationContext": {
        "in": "header",
        "name": "X-MediaHub-App-Id",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "IfNoneMatch": {
        "in": "header",
        "name": "If-None-Match",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "ObjectVersionId": {
        "in": "path",
        "name": "version_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
...
      "Range": {
        "in": "header",
        "name": "Range",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Permission denied"
      },
...
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Resource not found"
      },
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authentication required"
      },
      "Unavailable": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Dependency unavailable"
      },
...
    },
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "HmacAccessKey": {
        "description": "MediaHub-HMAC-SHA256 signed request with timestamp, nonce, body hash, and signed headers.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "MetricsBearer": {
        "description": "Dedicated deployment metrics bearer token.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "mediahub_session",
        "type": "apiKey"
      },
      "SignedMediaToken": {
        "description": "Short-lived token bound to one media revision and response policy.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      },
      "UploadCapability": {
        "description": "Short-lived capability bound to PUT and one upload session.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Docker Profile V1 API contract. Authenticated resource ownership is always derived from the session or API key, never from client-supplied owner IDs. Every response includes X-Request-Id; JSON errors additionally expose error.request_id.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "MediaHub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/object-versions/{version_id}/content": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationContext"
          },
          {
            "$ref": "#/components/parameters/ObjectVersionId"
          },
          {
            "$ref": "#/components/parameters/Range"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful response"
          },
          "206": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Partial content"
          },
          "304": {
            "description": "Not modified"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "416": {
            "description": "Range not satisfiable"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HmacAccessKey": []
          }
        ],
        "summary": "Read immutable ObjectVersion content"
      },
      "head": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationContext"
          },
          {
            "$ref": "#/components/parameters/ObjectVersionId"
          },
          {
            "$ref": "#/components/parameters/Range"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "206": {
            "description": "Partial content"
          },
          "304": {
            "description": "Not modified"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "416": {
            "description": "Range not satisfiable"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HmacAccessKey": []
          }
        ],
        "summary": "Read immutable ObjectVersion content headers"
      }
    },
...
  }
}